requirements.txt
torch>=2.0.0
transformers>=4.31.0
peft>=0.4.0
datasets>=2.13.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
tqdm>=4.65.0
PyPDF2>=3.0.0
rouge>=1.0.1
nltk>=3.8.1
llama-cpp-python>=0.2.0